% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareData.R
\name{rareData}
\alias{rareData}
\title{Table data colection}
\usage{
rareData(table)
}
\arguments{
\item{table}{containing the information necessary to calculate Rabinowitz forms of rarity}
}
\value{
The output is another table organised to be used in the function rarityForms
}
\description{
quantifies geographic range, abundance and habitat number for each species
}
