% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareForms.R
\name{rareForms}
\alias{rareForms}
\title{Seven forms of rarity}
\usage{
rareForms(data, percentage = 0.1, min_abund = 2, habitats = 1)
}
\arguments{
\item{data}{2B documented => is this correct? output provide by function rabiData}

\item{percentage}{2B documented.}

\item{min_abund}{2B documented.}

\item{habitats}{2B documented.}
}
\value{
The output is another table organised to be used in the function rarityForms
}
\description{
classifies the seven forms of rarity according to geographic range, abundance and habitat
}
