% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rareForms.R
\name{rareForms}
\alias{rareForms}
\title{Seven forms of rarity}
\usage{
rareForms(data, percentage = 0.1, min_abund = 2, habitats = 1)
}
\arguments{
\item{data}{output provide by function rabiData}

\item{percentage}{the percentage of localities below which the species will fulfil the first parametre to be considered rare. Default is 10%, but the user can change it.}

\item{min_abund}{the abundace  above which the species will fulfil the second parametre to be considered rare. Default is 2, but the user can change it.}

\item{habitats}{the number of habitats above which the species will fulfil the third parametre to be considered rare. Default is 1, but the user can change it.}
}
\value{
The output is the result table showing the name of each species in the first column and the form of rarity (or common) in the second column.
}
\description{
classifies the seven forms of rarity according to geographic range, abundance and habitat
}
