% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rarePlot.R
\name{rarePlot}
\alias{rarePlot}
\title{Local forms of rarity}
\usage{
rarePlot(table, rarity, locality)
}
\arguments{
\item{table}{containing the information of species occurrence per locality}

\item{rarity}{the output provided by the function rareForms}

\item{locality}{character, name of the target locality}
}
\value{
The function plots a pie chart showing the proportion of each rarity form for a target locality
}
\description{
plots a pie chart showing the shares of rarity forms in a specif locality
}
